supply_chain_app/requirements.txt
streamlit
pandas
duckdb
plotly
scikit-learn
google-generativeai
python-dotenv
